better loading msgs and bug fixing
